requirements.txt
firebase-admin
secure-smtplib
python-dotenv
fastapi
uvicorn
gunicorn
